Kingdoms creatable again
The previous requirements demanded that the character be a vassal to the current king. This caused issues if there was no king on the throne, as it checked the characters instead of the title. The checks now look to see if the primary title is de jure the next title up the chain. I'm not opposed to removing these, but I'll leave the original intent intact, keeping outsiders out.